+++ src/rsshub/mod.rs
@@ -24,8 +24,8 @@ impl RssHubClient {
     }
 
     /// Fetch youtube channel feed
-    pub async fn fetch_picuki(&self, account: &str) -> RssHubResult<Feed> {
-        let uri = format!("{}/picuki/profile/{}", self.url, account);
+    pub async fn fetch_instagram(&self, account: &str) -> RssHubResult<Feed> {
+        let uri = format!("{}/instagram/user/{}", self.url, account);
         let body = self.fetch_feed(&uri).await?;
         trace!("Got body {}", body);
         self.parse_feed(body)