+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.3.1](#031)
   - [0.3.0](#030)
   - [0.2.2](#022)
   - [0.2.1](#021)
@@ -9,6 +10,12 @@
   - [0.1.0](#010)
 
 ---
+
+## 0.3.1
+
+Released on 25/08/2022
+
+- Fetch instagram instead of picuki
 
 ## 0.3.0